Add right types to HOC
