+++ src/index.ts
@@ -1,9 +1,10 @@
 // Add a style tag to the document's head for the map's styling
 import injectCSS from './util/inject-css';
 import Map from './map';
-import Layer from './layer';
+import BaseLayer, { LayerCommonProps, Props as LayerProps } from './layer';
+import layerMouseTouchEvents, { EnhancedLayerProps } from './layer-events-hoc';
 import GeoJSONLayer from './geojson-layer';
-import Feature from './feature';
+import Feature, { Props } from './feature';
 import ZoomControl from './zoom-control';
 import Popup from './popup';
 import ScaleControl from './scale-control';
@@ -11,8 +12,19 @@ import Marker from './marker';
 import Source from './source';
 import Cluster from './cluster';
 import RotationControl from './rotation-control';
+import { Context } from './util/types';
 
 injectCSS(window);
+
+// Hack to get around import issue from external lib, see: https://github.com/Microsoft/TypeScript/issues/9944
+// TODO: Remove this hack once the above issue is fixed
+export type FeatureProps = Props;
+export type Context = Context;
+export type LayerCommonProps = LayerCommonProps;
+export type LayerProps = LayerProps;
+export type EnhancedLayerProps = EnhancedLayerProps;
+
+const Layer = layerMouseTouchEvents(BaseLayer);
 
 export {
   Feature,
